separating some of the code, utils and config defaults
